--- src/store/atom/cartItemQuantity.ts
@@ -1,6 +1,0 @@
-import { atomFamily } from "recoil";
-
-export const itemQuantityState = atomFamily<number, number>({
-  key: "itemQuantityState",
-  default: 0,
-});
